move handling labels logic into labels component and remove it from NoteForm
